修复删除选手BUG和导入选手BUG

+++ rate/rateserver/rate-service/src/main/java/org/sys/rate/service/admin/ParticipatesService.java
@@ -418,7 +418,9 @@ public class ParticipatesService {
 
     public Integer deleteParticipant(Integer groupID,Participates company) {
         int result = participatesMapper.delete(company);
-        studentMapper.deleteStudent(company.getStudentID());
+        if (participatesMapper.existPar(company.getStudentID()) == null){   // 该选手可能有其他活动，只有在选手表内无该选手时候，才能在student表删除
+            studentMapper.deleteStudent(company.getStudentID());
+        }
         if(result>0)
         {
             participatesMapper.subDisplaySequence(groupID,company.getDisplaySequence(),null);
